feat: bind data UI aiming bar

+++ Project/Assets/Needle.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Needle : MonoBehaviour
 {
-    public void IncreaseAngle(int delta)
+    public void IncreaseAngle(float delta)
     {
         transform.eulerAngles = new Vector3(0, 0, transform.eulerAngles.z - delta);
     }
